Add structure to separate the object mapPoints
For now, all mapPoints are saved in the same set, which is not sufficient
if we have more than one object.
Add the structure to keep the mapPoints related to each object. Need to adapt the
Tracking.

Change the YoloDetection parameters to test.
Change the object points (left, botton, top, right) logic.